Skeleton Site
This is the initial skeleton site.
The site is bootstrap based and used the "Flatly" theme
from http://bootswatch.com/flatly/

The majority of the content currently on the site is the standard
Lorem Ipsum which is used to validate the layout.

The homepage has some text along with some shortcode tests.
There is also some content in the what is programming page.

The feedback form is included along with the javascript
required to process it. This will only work if an
emailformgateway server process is running and listening
at the submit url of the form.