Shortcode for code figures, images and keyboard shortcuts
Added a shortcode for a code figure. This is an HTML5 figure
with a caption. The inner body of the shortcode should be the
source code of the program or a snippit of code.

The image shourtcode is a copy of the hugo img shortcode in the Hugo
docs.

The keyboard shortcodes come in three types. One for Windows and Linux
that uses the CTRL key + a key + and optional modifier. Or a MacOSX
version that uses the Command key + a key + an optional modifier.
The final type prints a keyboard shortcut in both styles to
avoid duplication.