Update the theme for the new lessons/lessonplans/teachernotes layout
A large update to the theme as a whole to prepare for the new
content layout structure.

The updates adds three new sections, lessons, lessonplans, and
teachernotes. The partials used to display each type of content
as both a single entry and a list have also been added.

Three new shortcodes have been added. One for abbreviations,
one for an expanding button - to create an expanding panel -
and one to create a figure that contains source code that can be
referenced later.