Added the new GopherCoders in Schools section
Added the new GopherCoders in Schools section to outline
how the approach can be used in both primary schools as
a code club and in secondary schools as part of the curriculum.

Updated the nav-bar to add the new option and added a two
new short codes for the quotes. Also included the mandelbrot
png resource