Add content to the feedback page
This quite a large update to the feedback page.
Superficially the update adds some content to explain what the page
is for and updates the form.

However there are a number of changes to the theme that allow the
form itself and the content above it to be separated. This has also
required a small change in the nav bar.