Add a summary section to all lessons
Added a summary, introduction "What You are Going to Learn" section
to all lessons.

The summaries must finish with the summary divider tag
<!--more-->
Doing so preserves the HTML formatting when the .Summary variable
is used in the templates.

The change means that formatting is now preserved in both the
lessons list page and the sidebar's featured lesson summary.