Update the littlegophers theme so that is works with Hugo >v0.17
Update the littlegophers theme so that it works again with newer
versions of hugo.

This approach is based on the new(ish) "baseof" html template under the
_default directory.

Renamed the theme to littlegophers2

Also added the bootstrap java script and CSS files under the static
directory. These seem to have been missing previously!