Fix shortcodes that include markdown and html
Fix all the shortcodes!

Hugo 0.55 changes the way shortcodes with markdown fenced with {{% ... %}} behaves.

The use of {{% ... %}} has been dropped in favour of {{< ... >}}

When using {{< ... >}} force the .Inner content of a shortcode
though the Markdown processor. This will ensure that the markdown is
rendered. And shortcode code that contained ".Inner" has been replaced
with ".Inner | markdownify"

Hugo v0.68 changes the default markdown processor from BlackFriday to
Goldmark. Goldmark is configured by default for security to ignore any
markdown that has embedded HTML - as the shortcodes do. This setting
needs to be changed to "unsafe" in the site wide condig.

These combined changes means the shortcodes render as perviously. So
panel* shortcodes can contain other short codes like highlight as well
as HTML divs and markdown.

l