Remove the "| markdownify" from th ".Inner" of a code figure
Code figures are supposed to contain code wrapped in a {{< highlight >}}
shortcode. There should not contain markdown that has to be parsed to HTML
embedded in the code and no code should be accidentally interpreted as markdown.

So we never want to pass the .Inner block into the markdown processor.