Change from linenos=table to linenos=inline
Change away from using a table for the line numbers in code to inline
line numbering.

Table based line numbering does not work with a responsive layout so
the numbering can be incorrect if the view is too narrow.

This change means that the code examples are no longer cut and paste friendly.